requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
opencv-python==4.9.0.80
mediapipe==0.10.9
mysql-connector-python==8.3.0
